requirements.txt
joblib==0.14.0
numpy==1.17.3
pandas==0.25.3
PySastrawi==1.2.0
python-dateutil==2.8.1
pytz==2019.3
scikit-learn==0.21.3
scipy==1.3.1
six==1.13.0
sklearn==0.0
